backend/requirements.txt
boto3>=1.34.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
